// src/interfaces/Species.ts
export default interface Species {
  name: string;
  classification: string;
  designation: string;
  average_height: string;
  skin_colors: string;
  hair_colors: string;
  eye_colors: string;
  average_lifespan: string;
  url: string;
  homeworld: string;
  language: string;
  people: Array<string>;
  films: Array<string>;
}